+++ src/decorators/types.ts
@@ -83,3 +83,5 @@ export type EnumValuesConfig<TEnum extends object> = Partial<
 >;
 
 export type MethodAndPropDecorator = PropertyDecorator & MethodDecorator;
+
+export type MethodPropClassDecorator = PropertyDecorator & MethodDecorator & ClassDecorator;