+++ src/metadata/metadata-storage.ts
@@ -3,6 +3,7 @@ import { NoExplicitTypeError } from "@/errors";
 import { type SchemaGeneratorOptions } from "@/schema/schema-generator";
 import { type ClassType } from "@/typings";
 import {
+  type AuthorizedClassMetadata,
   type AuthorizedMetadata,
   type BaseResolverMetadata,
   type ClassMetadata,
@@ -16,6 +17,7 @@ import {
   type ParamMetadata,
   type ResolverClassMetadata,
   type ResolverMetadata,
+  type ResolverMiddlewareMetadata,
   type SubscriptionResolverMetadata,
   type UnionMetadata,
   type UnionMetadataWithSymbol,
@@ -52,11 +54,15 @@ export class MetadataStorage {
 
   authorizedFields: AuthorizedMetadata[] = [];
 
+  authorizedResolver: AuthorizedClassMetadata[] = [];
+
   enums: EnumMetadata[] = [];
 
   unions: UnionMetadataWithSymbol[] = [];
 
   middlewares: MiddlewareMetadata[] = [];
+
+  resolverMiddlewares: ResolverMiddlewareMetadata[] = [];
 
   classDirectives: DirectiveClassMetadata[] = [];
 
@@ -110,6 +116,10 @@ export class MetadataStorage {
     this.authorizedFields.push(definition);
   }
 
+  collectAuthorizedResolverMetadata(definition: AuthorizedClassMetadata) {
+    this.authorizedResolver.push(definition);
+  }
+
   collectEnumMetadata(definition: EnumMetadata) {
     this.enums.push(definition);
   }
@@ -125,6 +135,10 @@ export class MetadataStorage {
 
   collectMiddlewareMetadata(definition: MiddlewareMetadata) {
     this.middlewares.push(definition);
+  }
+
+  collectResolverMiddlewareMetadata(definition: ResolverMiddlewareMetadata) {
+    this.resolverMiddlewares.push(definition);
   }
 
   collectResolverClassMetadata(definition: ResolverClassMetadata) {
@@ -190,9 +204,11 @@ export class MetadataStorage {
     this.argumentTypes = [];
     this.interfaceTypes = [];
     this.authorizedFields = [];
+    this.authorizedResolver = [];
     this.enums = [];
     this.unions = [];
     this.middlewares = [];
+    this.resolverMiddlewares = [];
     this.classDirectives = [];
     this.fieldDirectives = [];
     this.argumentDirectives = [];
@@ -213,12 +229,17 @@ export class MetadataStorage {
           field.params = this.params.filter(
             param => param.target === field.target && field.name === param.methodName,
           );
-          field.middlewares = mapMiddlewareMetadataToArray(
-            this.middlewares.filter(
-              middleware =>
-                middleware.target === field.target && middleware.fieldName === field.name,
+          field.middlewares = [
+            ...mapMiddlewareMetadataToArray(
+              this.resolverMiddlewares.filter(middleware => middleware.target === field.target),
             ),
-          );
+            ...mapMiddlewareMetadataToArray(
+              this.middlewares.filter(
+                middleware =>
+                  middleware.target === field.target && middleware.fieldName === field.name,
+              ),
+            ),
+          ];
           field.directives = this.fieldDirectives
             .filter(it => it.target === field.target && it.fieldName === field.name)
             .map(it => it.directive);
@@ -247,11 +268,17 @@ export class MetadataStorage {
         param => param.target === def.target && def.methodName === param.methodName,
       );
       def.roles = this.findFieldRoles(def.target, def.methodName);
-      def.middlewares = mapMiddlewareMetadataToArray(
-        this.middlewares.filter(
-          middleware => middleware.target === def.target && def.methodName === middleware.fieldName,
+      def.middlewares = [
+        ...mapMiddlewareMetadataToArray(
+          this.resolverMiddlewares.filter(middleware => middleware.target === def.target),
         ),
-      );
+        ...mapMiddlewareMetadataToArray(
+          this.middlewares.filter(
+            middleware =>
+              middleware.target === def.target && def.methodName === middleware.fieldName,
+          ),
+        ),
+      ];
       def.directives = this.fieldDirectives
         .filter(it => it.target === def.target && it.fieldName === def.methodName)
         .map(it => it.directive);
@@ -358,9 +385,10 @@ export class MetadataStorage {
   }
 
   private findFieldRoles(target: Function, fieldName: string): any[] | undefined {
-    const authorizedField = this.authorizedFields.find(
-      authField => authField.target === target && authField.fieldName === fieldName,
-    );
+    const authorizedField =
+      this.authorizedFields.find(
+        authField => authField.target === target && authField.fieldName === fieldName,
+      ) ?? this.authorizedResolver.find(authScope => authScope.target === target);
     if (!authorizedField) {
       return undefined;
     }