+++ src/shim.ts
@@ -45,7 +45,10 @@ export const Args: typeof src.Args = dummyDecorator;
 export const ArgsType: typeof src.ArgsType = dummyDecorator;
 export const Authorized: typeof src.Authorized = dummyDecorator;
 export const createParamDecorator: typeof src.createParamDecorator = dummyFn as any;
-export const createMethodDecorator: typeof src.createMethodDecorator = dummyFn as any;
+export const createMethodMiddlewareDecorator: typeof src.createMethodMiddlewareDecorator =
+  dummyFn as any;
+export const createResolverClassMiddlewareDecorator: typeof src.createResolverClassMiddlewareDecorator =
+  dummyFn as any;
 export const Ctx: typeof src.Ctx = dummyDecorator;
 export const Directive: typeof src.Directive = dummyDecorator;
 export const Extensions: typeof src.Extensions = dummyDecorator;