+++ tests/functional/authorization.ts
@@ -202,7 +202,7 @@ describe("Authorization", () => {
         normalQuery
       }`;
 
-      const result: any = await graphql({ schema, source: query });
+      const result: any = await graphql({ schema, source: query, contextValue: {} });
 
       expect(result.data!.normalQuery).toEqual(true);
     });
@@ -214,7 +214,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result: any = await graphql({ schema, source: query });
+      const result: any = await graphql({ schema, source: query, contextValue: {} });
 
       expect(result.data!.normalObjectQuery.normalField).toEqual("normalField");
     });
@@ -226,7 +226,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result: any = await graphql({ schema, source: query });
+      const result: any = await graphql({ schema, source: query, contextValue: {} });
 
       expect(result.data!.normalObjectQuery.normalResolvedField).toEqual("normalResolvedField");
     });
@@ -240,7 +240,26 @@ describe("Authorization", () => {
         authedQuery
       }`;
 
-      const result: any = await graphql({ schema: localSchema, source: query });
+      const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
+
+      expect(result.data).toBeNull();
+      expect(result.errors).toBeDefined();
+    });
+
+    it("should restrict access to authed resolver", async () => {
+      Authorized()(sampleResolver);
+      const localSchema = await buildSchema({
+        resolvers: [sampleResolver],
+        authChecker: () => false,
+      });
+
+      // clear AuthorizedClassMetadata for other tests
+      getMetadataStorage().authorizedResolver = [];
+      const query = `query {
+        normalQuery
+      }`;
+
+      const result = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data).toBeNull();
       expect(result.errors).toBeDefined();
@@ -256,7 +275,7 @@ describe("Authorization", () => {
         nullableAuthedQuery
       }`;
 
-      const result: any = await graphql({ schema: localSchema, source: query });
+      const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data!.nullableAuthedQuery).toBeNull();
       expect(result.errors).toBeUndefined();
@@ -271,7 +290,7 @@ describe("Authorization", () => {
         authedQuery
       }`;
 
-      const result: any = await graphql({ schema: localSchema, source: query });
+      const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data).toBeNull();
       expect(result.errors).toHaveLength(1);
@@ -289,7 +308,7 @@ describe("Authorization", () => {
         adminQuery
       }`;
 
-      const result: any = await graphql({ schema: localSchema, source: query });
+      const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data).toBeNull();
       expect(result.errors).toHaveLength(1);
@@ -323,7 +342,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result: any = await graphql({ schema: localSchema, source: query });
+      const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data).toBeNull();
     });
@@ -339,7 +358,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result: any = await graphql({ schema: localSchema, source: query });
+      const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data!.normalObjectQuery.nullableAuthedField).toBeNull();
     });
@@ -355,7 +374,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result: any = await graphql({ schema: localSchema, source: query });
+      const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data).toBeNull();
       expect(result.errors).toHaveLength(1);
@@ -375,7 +394,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result: any = await graphql({ schema: localSchema, source: query });
+      const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data).toBeNull();
       expect(result.errors).toHaveLength(1);
@@ -411,7 +430,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result: any = await graphql({ schema: localSchema, source: query });
+      const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data).toBeNull();
     });
@@ -427,7 +446,7 @@ describe("Authorization", () => {
         }
       }`;
 
-      const result: any = await graphql({ schema: localSchema, source: query });
+      const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data).toBeNull();
     });
@@ -464,6 +483,52 @@ describe("Authorization", () => {
       }`;
 
       const result: any = await graphql({ schema: localSchema, source: query, contextValue: {} });
+
+      expect(result.data!.adminOrRegularQuery).toEqual(false);
+      expect(authCheckerRoles).toEqual(["ADMIN", "REGULAR"]);
+    });
+
+    it("should pass roles from Resolver to `authChecker` when checking for access to handler", async () => {
+      let authCheckerRoles: string[] | undefined;
+      Authorized(["ADMIN", "REGULAR"])(sampleResolver);
+      const localSchema = await buildSchema({
+        resolvers: [sampleResolver],
+        authChecker: (_, roles) => {
+          authCheckerRoles = roles;
+          return true;
+        },
+      });
+
+      // clear AuthorizedClassMetadata for other tests
+      getMetadataStorage().authorizedResolver = [];
+      const query = `query {
+        normalQuery
+      }`;
+
+      const result = await graphql({ schema: localSchema, source: query, contextValue: {} });
+
+      expect(result.data!.normalQuery).toEqual(true);
+      expect(authCheckerRoles).toEqual(["ADMIN", "REGULAR"]);
+    });
+
+    it("should pass roles from Field rather than Resolver to `authChecker`", async () => {
+      let authCheckerRoles: string[] | undefined;
+      Authorized(["REGULAR"])(sampleResolver);
+      const localSchema = await buildSchema({
+        resolvers: [sampleResolver],
+        authChecker: (_, roles) => {
+          authCheckerRoles = roles;
+          return true;
+        },
+      });
+
+      // clear AuthorizedClassMetadata for other tests
+      getMetadataStorage().authorizedResolver = [];
+      const query = `query {
+        adminOrRegularQuery
+      }`;
+
+      const result = await graphql({ schema: localSchema, source: query, contextValue: {} });
 
       expect(result.data!.adminOrRegularQuery).toEqual(false);
       expect(authCheckerRoles).toEqual(["ADMIN", "REGULAR"]);